+++ src/main/java/com/gamja/edubox_team1/user/model/dto/UserRequestDTO.java
@@ -1,4 +1,4 @@
-package com.gamja.edubox_team1.model.dto;
+package com.gamja.edubox_team1.user.model.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.*;
