+++ src/main/java/com/gamja/edubox_team1/user/model/enums/UserRole.java
@@ -1,4 +1,4 @@
-package com.gamja.edubox_team1.model.enums;
+package com.gamja.edubox_team1.user.model.enums;
 
 public enum UserRole {
     STUDENT, CREATOR, ADMIN
